moar progress

+++ src/scene/workspacescene_opengl.h
@@ -102,7 +102,7 @@ private:
     void renderPart(const QRect &rect, const QRect &partRect, const QPoint &textureOffset, qreal devicePixelRatio, bool rotated = false);
     static const QMargins texturePadForPart(const QRect &rect, const QRect &partRect);
     void resizeTexture();
-    int toNativeSize(int size) const;
+    int toNativeSize(double size) const;
     std::unique_ptr<GLTexture> m_texture;
 };
